rm price reg 1.0
