+++ core/services/ocr2/plugins/ccip/internal/ccipdata/factory/price_registry.go
@@ -12,9 +12,7 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
 	ccipconfig "github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/config"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipcalc"
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipcommon"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_0_0"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_2_0"
 )
 
@@ -37,23 +35,6 @@ func initOrClosePriceRegistryReader(ctx context.Context, lggr logger.Logger, ver
 	}
 
 	contractType, version, err := versionFinder.TypeAndVersion(priceRegistryAddress, cl)
-	isV1_0_0 := ccipcommon.IsTxRevertError(err) || (contractType == ccipconfig.PriceRegistry && version.String() == ccipdata.V1_0_0)
-	if isV1_0_0 {
-		lggr.Infof("Assuming %v is 1.0.0 price registry, got %v", priceRegistryEvmAddr, err)
-		// Unfortunately the v1 price registry doesn't have a method to get the version so assume if it reverts its v1.
-		pr, err2 := v1_0_0.NewPriceRegistry(lggr, priceRegistryEvmAddr, lp, cl, registerFilters)
-		if err2 != nil {
-			return nil, err2
-		}
-		if closeReader {
-			return nil, pr.Close()
-		}
-		return pr, nil
-	}
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to read type and version")
-	}
-
 	if contractType != ccipconfig.PriceRegistry {
 		return nil, errors.Errorf("expected %v got %v", ccipconfig.PriceRegistry, contractType)
 	}
