+++ core/services/ocr2/plugins/ccip/internal/ccipdata/factory/price_registry_test.go
@@ -10,6 +10,7 @@ import (
 	cciptypes "github.com/smartcontractkit/chainlink-common/pkg/types/ccip"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
+
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
 	mocks2 "github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller/mocks"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/utils"
@@ -21,7 +22,7 @@ import (
 func TestPriceRegistry(t *testing.T) {
 	ctx := testutils.Context(t)
 
-	for _, versionStr := range []string{ccipdata.V1_0_0, ccipdata.V1_2_0} {
+	for _, versionStr := range []string{ccipdata.V1_2_0} {
 		lggr := logger.Test(t)
 		addr := cciptypes.Address(utils.RandomAddress().String())
 		lp := mocks2.NewLogPoller(t)
